Remove sound files impossible to redistribute
The ringtone sound file wasn’t licensed for redistribution as far as I
know, this prevents JSXC from being allowed to redistribute it without
written permission from the original author.

If you want to keep using this sound file you could try to contact the
author and see if they agree with you redistributing it, in the meantime
it’s safer both for the project, for its users and for its downstream
distributors to only use free sound files.

The new message sound file wasn’t even sourced, this makes it even more
problematic as the author can’t be contacted at all.