feat: add XEP-0308 Last Message Correction
This PR adds the ability to correct messages. Users have to long press on a message bubble to
enter the edit mode. Corrected messages are marked with a pencil icon.
There is an mouseover tooltip (about the pencil icon) to view the corrected messages.
It also supports mam messages.

fix: sync conversations after login

This PR will fix an issue of loading messages from mam:
Case: if you are in an conversation with jsxc and logout and then continue the conversation with an other client or another instance of jsxc with other storage and then relogin with the instance mentioned before, then jsxc wont load the newly written messages from mam.
This PR will add the ability to sync all contacts after login and takes the last message timestamp as the beginning for them mam request.
It is only needed one time after login.

fix: add checks for xep0421 on muc messages to validate message corrections in some way if the muc is anonymous