fix: only allow correction of last message
but added new option "allowAllMessagesCorrection" to allow correction of all messages

fix: If the same message is to be corrected several times, the id of the original message is used in each case (e.g. If a message of id 1 is corrected by a message of id 2, a subsequent correction should correct 1, not 2).