feat: show geodata from XEP-0080 message events in roster
This commit adds the ability to show the contacts position if it will be
sent through message events (XEP-0080). An indicator will be added on the
contacts avatar. A mouseover will show the address (if it could be reversed geocoded).
Otherwhise a click on the indicator will open openstreetmap and show the position on a map.